Move event details to separate component. Add tests
