+++ client/app/components/calendar/event/DeleteConfirmation.jsx
@@ -2,10 +2,8 @@ import React from 'react';
 import { Modal, Button } from 'react-bootstrap';
 import EventSchema from 'schemas/EventSchema';
 import { eventTimeString } from 'helpers/DateHelper';
-import moment from 'moment';
-
-const TIME_FORMAT = 'HH:mm';
-const DATE_FORMAT = 'Mo MMM YYYY';
+import { formatDate } from 'helpers/DateHelper';
+import EventDetails from './EventDetails';
 
 const DeleteConfirmation = ({ onHide, show, onCancel, onConfirm, event }) => (
   <Modal show={show}
@@ -15,20 +13,8 @@ const DeleteConfirmation = ({ onHide, show, onCancel, onConfirm, event }) => (
       Are you sure you want to delete this event?
     </Modal.Header>
     <Modal.Body className="delete-confirmation-modal modal-body">
-      <h1 className="summary">{event.summary}</h1>
-      <p>
-        <em>on </em>
-        <strong>{moment(event.start.date_time).format(DATE_FORMAT)} </strong>
-        ({eventTimeString(event, TIME_FORMAT)})
-      </p>
-      <p>
-        <em>in </em>
-        {event.conference_room.title}
-      </p>
-      <p>
-        <em>by </em>
-        {event.creator.email}
-      </p>
+      <EventDetails event={event}
+                    timeFormat="MM/DD HH:mm" />
     </Modal.Body>
     <Modal.Footer className="delete-confirmation-modal">
       <Button onClick={onCancel}
