+++ client/app/components/calendar/event/Event.jsx
@@ -1,7 +1,7 @@
 import React, { PropTypes } from 'react';
 import EventSchema from 'schemas/EventSchema';
 import EventDestroyer from './EventDestroyer';
-import { eventTimeString } from 'helpers/DateHelper';
+import EventDetails from './EventDetails';
 import './event.scss';
 
 export default class Event extends React.Component {
@@ -23,16 +23,7 @@ export default class Event extends React.Component {
         <EventDestroyer onDelete={this.props.onDelete}
                         disabled={!creator.self}
                         event={this.props.event} />
-        <div className="event-time">{eventTimeString(this.props.event, this.props.timeFormat)}</div>
-        <div className="event-name">{event.summary}</div>
-        <div className="event-user">
-          <small>by&nbsp;</small>
-          {creator.display_name || creator.email}
-        </div>
-        <div className="event-location">
-          <small>in&nbsp;</small>
-          {event.conference_room.title}
-        </div>
+        <EventDetails event={event} timeFormat={this.props.timeFormat} />
       </div>
     );
   }
